Vitals: Nightscout: use decimal for values
Apparently, some of the numeric values supplied by Nightscout are not
integral. Deal with that by using the decimal datatype.

+++ Plugins/Vitals/Nightscout/NightscoutMBGEntry.cs
@@ -26,7 +26,7 @@ namespace SharpIrcBot.Plugins.Vitals.Nightscout
         public DateTimeOffset Timestamp { get; set; }
 
         [JsonProperty("mbg")]
-        public int MBG { get; set; }
+        public decimal MBG { get; set; }
         
         [JsonProperty("sysTime")]
         public string SystemTimeString
